refactor: decomposition test
